feat: 파라미터를 통한 파이어베이스 방 정보 가져오기
