+++ src/views/index.js
@@ -1,4 +1,5 @@
 export { default as LiarGameView } from './LiarGameView';
-export { default as MainView } from './MainView';
+export { default as WaitingRoomView } from './WaitingRoomView';
 export { default as WordGameView } from './WordGameView';
 export { default as RouletteGameView } from './RouletteGameView';
+export { default as MainView } from './MainView';
